python/requirements.txt
pyserial==3.5
pillow==9.2.0
black==23.12.1
flake8==6.1.0
